Changing references of unsigned ints to size_t, or to use of iterators. Also fixing one bug in Object2D::getNumDistinctX, where we weren't looking at all scans. This should fix ticket #63.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@623 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Cubes/Merger.cc
@@ -70,7 +70,7 @@ namespace duchamp
       // Do growth stuff
       if(this->par.getFlagGrowth()) {
 	vector <Detection> newList(currentList.size());
-	for(unsigned int i=0;i<currentList.size();i++){
+	for(size_t i=0;i<currentList.size();i++){
 	  std::cout.setf(std::ios::right);
 	  std::cout << "Growing: " << std::setw(6) << i+1 << "/";	   
 	  std::cout.unsetf(std::ios::right);
@@ -97,7 +97,7 @@ namespace duchamp
 
       //     *this->objectList = currentList;
       this->objectList->resize(currentList.size());
-      for(unsigned int i=0;i<currentList.size();i++)
+      for(size_t i=0;i<currentList.size();i++)
 	this->objectList->at(i) = currentList[i];
     
       currentList.clear();
@@ -129,7 +129,7 @@ namespace duchamp
       bool isVerb = par.isVerbose();
       vector <Detection>::iterator iter;
 
-      unsigned int counter=0, compCounter;
+      size_t counter=0, compCounter;
       while( counter < (objList.size()-1) ){
 	if(isVerb){
 	  std::cout.setf(std::ios::right);
