+++ src/Cubes/VOTable.cc
@@ -63,7 +63,7 @@ namespace duchamp
     /// \return String with fixed units
 
     std::string newstring;
-    for(unsigned int i=0;i<oldstring.size();i++){
+    for(size_t i=0;i<oldstring.size();i++){
       if((oldstring[i]!='[')&&(oldstring[i]!=']')){
 	if(oldstring[i]==' ') newstring += '.';
 	else newstring += oldstring[i];
