+++ src/Detection/sorting.cc
@@ -122,7 +122,7 @@ namespace duchamp
     /// case it is returned unaltered.
 
     bool isGood = true;
-    for(unsigned int i=0;i<inputList.size();i++) isGood = isGood && inputList[i].isWCS();
+    for(size_t i=0;i<inputList.size();i++) isGood = isGood && inputList[i].isWCS();
 
     if(isGood){
 
@@ -182,7 +182,7 @@ namespace duchamp
 
     bool isGood = true;
     if(parameter!="zvalue"){
-      for(unsigned int i=0;i<inputList.size();i++) isGood = isGood && inputList[i].isWCS();
+      for(size_t i=0;i<inputList.size();i++) isGood = isGood && inputList[i].isWCS();
     }
 
     if(isGood){
